# loms/pkg/api/order/v1/order.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Order API",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Order"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/order/cancel": {
      "post": {
        "operationId": "Order_Cancel",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CancelRequest"
            }
          }
        ],
        "tags": [
          "Order"
        ]
      }
    },
    "/v1/order/create": {
      "post": {
        "operationId": "Order_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateRequest"
            }
          }
        ],
        "tags": [
          "Order"
        ]
      }
    },
    "/v1/order/info/{order_id}": {
      "get": {
        "operationId": "Order_Info",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1InfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Order"
        ]
      }
    },
    "/v1/order/pay": {
      "post": {
        "operationId": "Order_Pay",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1PayRequest"
            }
          }
        ],
        "tags": [
          "Order"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CancelRequest": {
      "type": "object",
      "properties": {
        "order_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1CreateRequest": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1CreateRequestItem"
          }
        }
      }
    },
    "v1CreateRequestItem": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "count": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1CreateResponse": {
      "type": "object",
      "properties": {
        "order_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1InfoResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1InfoResponseItem"
          }
        }
      }
    },
    "v1InfoResponseItem": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "count": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1PayRequest": {
      "type": "object",
      "properties": {
        "order_id": {
          "type": "string",
          "format": "int64"
        }
      }
    }
  },
  "securityDefinitions": {
    "x-auth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  }
}